Fix bugs

+++ src/dataset/collate.py
@@ -21,7 +21,7 @@ def collate_fn(dataset_items: list[dict]) -> dict[Union[torch.Tensor, list]]:
 
     result_batch["images"] = torch.stack([item["image"] for item in dataset_items])
     if "label" in dataset_items[0].keys():
-        result_batch["labels"] = torch.IntTensor(
+        result_batch["labels"] = torch.LongTensor(
             [item["label"] for item in dataset_items]
         )
